7. Разделяй и властвуй: раскидал фаил utils.js на две части (общее и отрисовка), модернезировал функцию render для работы с элементами
